app/backend/requirements.txt
fastapi>=0.115.4
uvicorn[standard]>=0.32.1
peewee>=3.17.8
pymysql>=1.1.1
pydantic>=2.10.3
python-multipart>=0.0.18
python-dotenv>=1.0.0
cryptography>=41.0.4